+++ src/gameboy/memory/memory.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string>
+#include <stdint.h>
 
 using namespace std;
 
@@ -46,7 +47,7 @@ class Memory {
         ~Memory();
 
         // Memory read functions
-        char& fetch8(int &address); // Fetch 8-bit value from memory
+        char& fetch8(uint16_t &address); // Fetch 8-bit value from memory
         // int& fetch16(int &address); // Fetch 16-bit value from memory
 
         // ROM load functions
